src/requirements.txt
fastapi
uvicorn
pillow
requests
boto3
python-dotenv
redis
cachetools